move Image to components/core
