--- src/components/Image/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "Image.js",
-  "version": "0.0.0",
-  "main": "./Image.js"
-}
