+++ src/components/ProfileImage/ProfileImage.js
@@ -21,7 +21,7 @@ import * as React from 'react';
 import styled, { css } from 'styled-components/native';
 
 // Components
-import Image from 'components/Image';
+import Image from 'components/core/Image';
 
 // Utils
 import { getIdenticonImageUrl } from 'utils/images';
