+++ src/components/SWActivationModal/SWActivationModal.js
@@ -33,7 +33,7 @@ import { deploySmartWalletAction, estimateSmartWalletDeploymentAction } from 'ac
 // components
 import SlideModal from 'components/Modals/SlideModal';
 import Button from 'components/legacy/Button';
-import Image from 'components/Image';
+import Image from 'components/core/Image';
 import { MediumText, BaseText } from 'components/legacy/Typography';
 import { Spacing } from 'components/legacy/Layout';
 import FeeLabelToggle from 'components/FeeLabelToggle';
