+++ src/components/legacy/EventDetails/EventDetails.js
@@ -32,7 +32,7 @@ import t from 'translations/translate';
 // components
 import { BaseText, MediumText } from 'components/legacy/Typography';
 import Icon from 'components/legacy/Icon';
-import Image from 'components/Image';
+import Image from 'components/core/Image';
 import TankAssetBalance from 'components/TankAssetBalance';
 import ReceiveModal from 'screens/Asset/ReceiveModal';
 import SWActivationModal from 'components/SWActivationModal';
