+++ src/components/legacy/ValueInput/ValueInputHeader.js
@@ -29,7 +29,7 @@ import { useTranslation } from 'translations/translate';
 
 // components
 import Icon from 'components/legacy/Icon';
-import Image from 'components/Image';
+import Image from 'components/core/Image';
 import { BaseText, MediumText } from 'components/legacy/Typography';
 import { Spacing } from 'components/legacy/Layout';
