+++ src/screens/BiometricsPrompt/BiometricsPrompt.js
@@ -33,7 +33,7 @@ import { beginOnboardingAction } from 'actions/onboardingActions';
 import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
 import { MediumText } from 'components/legacy/Typography';
 import Button from 'components/legacy/Button';
-import Image from 'components/Image';
+import Image from 'components/core/Image';
 import Toast from 'components/Toast';
 
 // utils
