+++ src/screens/KeyBasedAssetTransfer/KeyBasedAssetTransferConfirm.js
@@ -31,7 +31,7 @@ import BalanceView from 'components/BalanceView';
 import Button from 'components/legacy/Button';
 import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
 import FeeTable from 'components/display/FeeTable';
-import Image from 'components/Image';
+import Image from 'components/core/Image';
 import Spinner from 'components/Spinner';
 import Text from 'components/core/Text';
