+++ src/screens/LiquidityPools/LiquidityPools.js
@@ -27,7 +27,7 @@ import t from 'translations/translate';
 // components
 import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
 import Table, { TableRow, TableLabel } from 'components/legacy/Table';
-import Image from 'components/Image';
+import Image from 'components/core/Image';
 import { Spacing, ScrollWrapper } from 'components/legacy/Layout';
 import { BaseText } from 'components/legacy/Typography';
 import Tabs from 'components/legacy/Tabs';
