+++ src/screens/PillarNetwork/PillarNetworkIntro.js
@@ -28,7 +28,7 @@ import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
 import { ScrollWrapper } from 'components/legacy/Layout';
 import { BoldText, MediumText } from 'components/legacy/Typography';
 import Icon from 'components/legacy/Icon';
-import Image from 'components/Image';
+import Image from 'components/core/Image';
 import Button from 'components/legacy/Button';
 import { LabelBadge } from 'components/LabelBadge';
