+++ src/screens/WalletConnect/CallRequest/SignatureRequestContent.js
@@ -24,7 +24,7 @@ import styled from 'styled-components/native';
 
 // Components
 import Button from 'components/core/Button';
-import Image from 'components/Image';
+import Image from 'components/core/Image';
 import Text from 'components/core/Text';
 
 // Constants
