+++ src/screens/WalletMigrationArchanova/Review/MigrationSummary.js
@@ -27,7 +27,7 @@ import { useTranslationWithPrefix } from 'translations/translate';
 import * as Table from 'components/layout/Table';
 import BalanceView from 'components/BalanceView';
 import FeeTable from 'components/display/FeeTable';
-import Image from 'components/Image';
+import Image from 'components/core/Image';
 import Text from 'components/core/Text';
 
 // Constants
